+++ changelog.md
@@ -5,6 +5,12 @@ All notable changes done in ODS library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://github.com/Orange-OpenSource/ods-android/compare/0.15.1...develop)
+
+### Changed
+
+- \[App\] Change buttons name ([#647](https://github.com/Orange-OpenSource/ods-android/issues/647))
+
 ## [0.15.1](https://github.com/Orange-OpenSource/ods-android/compare/0.15.0...0.15.1) - 2023-09-28
 
 ### Changed